requirements.txt
certifi==2015.4.28
cssselect==0.9.2
Django==1.8.2
django-debug-toolbar==1.3.0
django-debug-toolbar-extra==0.0.2
django-debug-toolbar-template-timings==0.6.4
django-formtools==1.0
django-haystack==2.4.0
django-honeypot==0.4.0
django-jquery==1.9.1
django-static-sitemaps==2.1.7
django-wysiwyg-redactor==0.4.7
elasticsearch==1.6.0
elasticstack==0.1.1
ipdb==0.8
ipython==3.1.0
lxml==3.3.5
Pillow==2.8.1
psycopg2==2.6
pyelasticsearch==1.4
redis==2.10.3
requests==2.11.1
simplejson==3.7.3
six==1.9.0
sqlparse==0.1.15
urllib3==1.10.4
